# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "LTI Canvas Devops Backend",
    "description": "LTI Canvas Devops API Documentation"
  },
  "schemes": [
    "http",
    "https"
  ],
  "host": "localhost:3000",
  "basePath": "/api",
...
  "paths": {
...
    "/login/register_user": {
      "post": {
        "tags": [
          "User Login Register"
        ],
        "summary": "Register user in Canvas DevOps portal",
        "description": "Returns status of user registration",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "User details needed for registration",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "integer",
                  "format": "int64",
                  "example": "10645213",
                  "description": "PS number of the user"
                },
                "user_name": {
                  "type": "string",
                  "example": "Darshan Bhojani"
                },
                "user_mail": {
                  "type": "string",
                  "example": "darshan.bhojani@lntinfotech.com"
                },
                "user_password": {
                  "type": "string",
                  "example": "qe45WG53"
                },
                "security_question": {
                  "type": "string",
                  "example": "What is your birth place?"
                },
                "security_answer": {
                  "type": "string",
                  "example": "Mumbai"
                },
                "user_allocation": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "role_name": {
                        "type": "string",
                        "description": "Role assigned to the team member",
                        "enum": [
                          "Project Manager",
                          "Developer",
                          "Scrum Master",
                          "Architect"
                        ]
                      },
                      "application_name": {
                        "type": "string"
                      },
                      "application_key": {
                        "type": "string",
                        "description": "Key of the application assigned in the portal"
                      },
                      "isAuthorized": {
                        "type": "boolean",
                        "description": "Whether user has edit access"
                      },
                      "pipelines": {
                        "type": "array",
                        "items": {
                          "properties": {
                            "pipeline_name": {
                              "type": "string"
                            },
                            "pipeline_key": {
                              "type": "string",
                              "description": "Key of the pipeline assigned in the portal"
                            }
                          }
                        }
                      }
                    }
                  },
                  "example": [],
                  "description": "user Allocation always empty array when user is registering"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful registration of user",
            "schema": {
              "properties": {
                "status": {
                  "type": "string",
                  "example": "User registration successful"
                }
              }
            }
          },
          "500": {
            "description": "If user already exists or error while adding user to database",
            "schema": {
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Error while registering user"
                }
              }
            }
          }
        }
      }
    },
    "/login/login_user": {
      "post": {
        "tags": [
          "User Login Register"
        ],
        "summary": "Login user in Canvas DevOps portal",
        "description": "Returns status of user login",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "User info required for login",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_name": {
                  "type": "string",
                  "example": "Darshan Bhojani"
                },
                "user_password": {
                  "type": "string",
                  "example": "qe45WG53"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful authentication of user",
            "schema": {
              "properties": {
                "status": {
                  "type": "string",
                  "example": "User authentication successful"
                },
                "no_of_apps": {
                  "type": "number",
                  "example": 3
                }
              }
            }
          },
          "500": {
            "description": "User authentication failed",
            "schema": {
              "properties": {
                "status": {
                  "type": "string",
                  "example": "User authentication failed"
                }
              }
            }
          }
        }
      }
    },
    "/home/application_list": {
      "get": {
        "tags": [
          "Home"
        ],
        "summary": "Get application list for logged in user",
        "description": "Returns list of application that user is assigned to",
        "parameters": [
          {
            "name": "user_name",
            "in": "query",
            "description": "Logged in user's user name",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of applications",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "500": {
            "description": "Error while fetching user applications",
            "schema": {
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Error while fetching user applications"
                }
              }
            }
          }
        }
      }
    },
    "/home/add_application": {
      "post": {
        "tags": [
          "Home"
        ],
        "summary": "Add application for user",
        "description": "Returns status of added application",
        "parameters": [
          {
            "in": "body",
            "name": "application",
            "description": "Application details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_name": {
                  "type": "string",
                  "example": "Darshan Bhojani"
                },
                "application": {
                  "type": "object",
                  "properties": {
                    "application_name": {
                      "type": "string",
                      "example": "Citi Bank"
                    },
                    "application_type": {
                      "type": "string",
                      "example": "AD"
                    },
                    "application_desc": {
                      "type": "string",
                      "example": "Application to handle Citi Banks transaction"
                    },
                    "project_manager": {
                      "type": "string",
                      "example": "Neha Dhondhe"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status of added application",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Successfully added application"
                }
              }
            }
          },
          "500": {
            "description": "If application already exists or error while adding application to database",
            "schema": {
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Error while adding application"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Application": {
      "type": "object",
      "properties": {
        "application_name": {
          "type": "string",
          "example": "Citi Bank"
        },
        "application_type": {
          "type": "string",
          "example": "AD"
        },
        "application_desc": {
          "type": "string",
          "example": "Application to handle Citi Banks transaction"
        },
        "application_health": {
          "type": "string",
          "example": "AT RISK"
        },
        "pipelines": {
          "type": "number",
          "example": 5,
          "description": "Number of pipelines in the application"
        },
        "created_at": {
          "type": "string",
          "example": "August 12, 2020"
        },
        "project_manager": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "example": "Rahul Mandhane"
            },
            "id": {
              "type": "number",
              "example": 10669771
            }
          }
        }
      }
    },
...
  }
}